// angular-reactive-form/src/app/shared/schemas/date-picker-options.ts
export class DatePickerOptions {
    singleDatePicker: boolean;
    showDropdowns: boolean;
    autoApply: boolean;
    drops: string;
    timePicker: boolean;
    autoUpdateInput: boolean;
    placeholder: string;
    minDate: object;
}

export class DatePickerOptionsAllDates {
    singleDatePicker: boolean;
    showDropdowns: boolean;
    autoApply: boolean;
    drops: string;
    timePicker: boolean;
    autoUpdateInput: boolean;
    placeholder: string;
}